pyproject.toml
[project]
name = "webbsite"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "flask",
    "psycopg2-binary",
    "gunicorn",
    "python-dotenv",
]
